Adding Euro Truck Simulator 2 keyboard actions.
I've checked this on ETS2 running on Windows under Steam. I did not try
on other systems, but I believe they are the same.

The official manual has an incomplete list:
https://steamcdn-a.akamaihd.net/steam/apps/227300/manuals/ETS2_manual_en.pdf

The complete list is either from third-party websites or from within the
game itself.

This site contains a neat layout:
http://isometricland.net/keyboard/keyboardchart.php?gam=115&sty=5&lay=1

It is worth noting that:

* There are some actions with 2 assigned keys. To work around
  ShorcutMapper behavior, I added a space character to the second
  binding.

* There are some actions without any assigned key. I've left those in
  the JSON file, but assigned to an empty string.

* There is one action using the right mouse button, but it is not yet
  supported.